+++ platform/ios/src/MGLMapViewDelegate.h
@@ -6,6 +6,7 @@
 NS_ASSUME_NONNULL_BEGIN
 
 @class MGLMapView;
+@class MGLUserLocationAnnotationViewStyle;
 
 /**
  The `MGLMapViewDelegate` protocol defines a set of optional methods that you
@@ -312,6 +313,17 @@ NS_ASSUME_NONNULL_BEGIN
  @param mapView The map view that is tracking the user’s location.
  */
 - (void)mapViewDidStopLocatingUser:(MGLMapView *)mapView;
+
+/**
+ Asks the delegate styling options for each default user location annotation view.
+ 
+ This method is called many times during gesturing, so you should avoid performing
+ complex or performance-intensive tasks in your implementation.
+ 
+ @param mapView The map view that is tracking the user’s location.
+ */
+- (MGLUserLocationAnnotationViewStyle *)mapViewStyleForDefaultUserLocationAnnotationView:(MGLMapView *)mapView NS_SWIFT_NAME(mapView(styleForDefaultUserLocationAnnotationView:));
+
 
 /**
  Tells the delegate that the location of the user was updated.